Organise transaction template in a separate file
